% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_row.R
\name{add_row_server}
\alias{add_row_server}
\title{Server logic to add a row}
\usage{
add_row_server(id, ppg)
}
\arguments{
\item{id}{Character vector of length 1; the ID for
this module.}

\item{ppg}{Reactive dataframe (tibble) of PPG data}
}
\value{
Server logic
}
\description{
Internal function
}
