% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_row.R
\name{add_row_ui}
\alias{add_row_ui}
\title{UI for adding a row}
\usage{
add_row_ui(id)
}
\arguments{
\item{id}{Character vector of length 1; the ID for
this module.}
}
\value{
UI
}
\description{
Internal function
}
