% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipni_authors}
\alias{ipni_authors}
\title{IPNI authors}
\format{
An object of class \code{character} of length 34605.
}
\usage{
ipni_authors
}
\description{
Taxonomic author names formatting according to the International Plant Names
Index (IPNI). Data obtained from the World Checklist of Vascular Plants
(WCVP) under the CC-BY license.
}
\details{
IPNI (2024). International Plant Names Index. Published on the Internet
http://www.ipni.org, The Royal Botanic Gardens, Kew, Harvard University
Herbaria & Libraries and Australian National Herbarium.
}
\keyword{datasets}
