% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{linkize_url_col}
\alias{linkize_url_col}
\title{Turn URLs into hyperlinks}
\usage{
linkize_url_col(url)
}
\arguments{
\item{url}{Vector of URLs}
}
\value{
Vector with URLs as hyperlinks
}
\description{
Internal function
}
