% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{linkize_urls}
\alias{linkize_urls}
\title{Turn URLs into hyperlinks in the ppg dataframe}
\usage{
linkize_urls(ppg, url_cols)
}
\arguments{
\item{ppg}{PPG datafame}

\item{url_cols}{Names of columns with URLs}
}
\value{
ppg with URL columns formatted as links
}
\description{
Internal function
}
