% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{ppg_app}
\alias{ppg_app}
\title{PPG app}
\usage{
ppg_app()
}
\value{
An object that represents the app. This function is normally called
for its side-effect of starting a Shiny app that can be accessed in a web
browser window.
}
\description{
Run a Shiny app to view and edit the Pteridophyte Phylogeny Group (PPG)
taxonomic database.
}
\examples{
if (interactive()) {
  ppg_app()
}
}
