% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppg_version}
\alias{ppg_version}
\title{PPG version}
\format{
An object of class \code{character} of length 1.
}
\usage{
ppg_version
}
\description{
Version of the PPG dataset used in ppg_full
}
\keyword{datasets}
