% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pterido_higher_names}
\alias{pterido_higher_names}
\title{Higher names of pteridophytes}
\format{
An object of class \code{character} of length 1420.
}
\usage{
pterido_higher_names
}
\description{
Names at the genus level and above of pteridophytes. Includes synonyms.
Does not include authors. Only meant to be used to verify spelling.
}
\keyword{datasets}
