% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pterido_sp_epithets}
\alias{pterido_sp_epithets}
\title{Specific epithets of pteridophytes}
\format{
An object of class \code{character} of length 15366.
}
\usage{
pterido_sp_epithets
}
\description{
Specific epithets of pteridophytes. Includes synonyms.
Does not include authors. Only meant to be used to verify spelling.
}
\keyword{datasets}
