% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{str_to_random}
\alias{str_to_random}
\title{Randomly capitalize the letters of a string}
\usage{
str_to_random(x, n_cap = NULL)
}
\arguments{
\item{x}{Character vector}

\item{n_cap}{Number of characters to capitalize. Default will capitalize
half (rounded down)}
}
\description{
Vectorized
}
