% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppg_small}
\alias{ppg_small}
\title{PPG taxonomic data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3163 rows and 11 columns.
}
\usage{
ppg_small
}
\description{
PPG data for testing this app
}
\keyword{datasets}
