% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppg_small}
\alias{ppg_small}
\title{PPG taxonomic data}
\format{
An object of class \code{data.frame} with 3232 rows and 19 columns.
}
\usage{
ppg_small
}
\description{
PPG data for testing this app
}
\keyword{datasets}
